requirements.txt
igdb_api_v4==0.2.0
joblib==1.2.0
pandas==1.3.5
rapidfuzz==3.2.0
requests==2.31.0
python-dotenv==1.0.0
sqlean.py==0.21.8.4
click==8.1.7
rich==13.6.0
